src/test/java/valueSets/WirkstofftypTest.java
package valueSets;

import org.junit.jupiter.api.Test;
import valueSet.CodeDto;
import valueSet.CodeUtil;

import java.io.FileNotFoundException;
import java.util.List;

import static org.junit.jupiter.api.Assertions.assertEquals;
import static valueSet.CodeUtil.assertSimpleSystem;
import static valueSet.CodeUtil.assertValidCodes;

class WirkstofftypTest {
  @Test
  void testSystem() {
    String expectedSystem =
        "https://www.medizininformatik-initiative.de/fhir/core/modul-medikation/CodeSystem/wirkstofftyp";
    assertSimpleSystem(expectedSystem, Wirkstofftyp.values());
  }

  @Test
  void testAllValidCodes() throws FileNotFoundException {
    List<CodeDto> codes = CodeUtil.get("WirkstofftypCodes.json");
    assertValidCodes(codes, Wirkstofftyp::fromCode);
    assertEquals(codes.size(), Wirkstofftyp.values().length);
  }
}
